+++ rsass/tests/spec/core_functions/color/grayscale.rs
@@ -164,7 +164,6 @@ fn with_calc() {
     );
 }
 #[test]
-#[ignore] // unexepected error
 fn with_css_var() {
     assert_eq!(
         runner().ok("a {b: grayscale(var(--c))}\n"),
@@ -174,7 +173,6 @@ fn with_css_var() {
     );
 }
 #[test]
-#[ignore] // unexepected error
 fn with_unquoted_calc() {
     assert_eq!(
         runner().ok("a {b: grayscale(unquote(\'calc(1)\'))}\n"),