+++ rsass/tests/spec/core_functions/color/saturate.rs
@@ -194,7 +194,6 @@ mod one_arg {
         );
     }
     #[test]
-    #[ignore] // unexepected error
     fn with_css_var() {
         assert_eq!(
             runner().ok("a {b: saturate(var(--c))}\n"),
@@ -214,7 +213,6 @@ mod one_arg {
         );
     }
     #[test]
-    #[ignore] // unexepected error
     fn with_unquoted_calc() {
         assert_eq!(
             runner().ok("a {b: saturate(unquote(\'calc(1)\'))}\n"),